refactor: sizes in language constructors
Constructors of languages used Size< for subtree size constraints so far. I
refactored these constraints to use regular sizes instead and to instead produce
an expression being one larger (↑ e). I did this following the advice of gallais from the zulip agda chat to not mix both size constructors.
He said ↑ is good for inductive data while Size< is good for co-inductive data (or codata).
This commit witnesses that indeed some size juggling efforts could be simplified but to a lesser
extent that I hoped for.