Apply the default arguments to the nixpkgs function
This fixes the error
```
error: value is a function while a set was expected

       at /home/benjamin/studium/VariantSync/AgdaCCnOC/default.nix:10:1:

            9| { pkgs ? import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/refs/tags/23.05.tar.gz") }:
           10| pkgs.agdaPackages.mkDerivation {
             | ^
           11|   version = "1.0";
```
which indicates that `pkgs` is a function although it should be a set
(the `.` operator requires a set on its left hand side).

Now, `nix-shell` actually works but doesn't provide all necessary
dependencies (the standard library is still commented).