Combine the restricted config and the selection language
Note that I reused the letter `S` out of laziness. This should be
changed if we go ahead with this approach.

A notable feature that is missing in this new approach is the ability to
evaluate the configuration. It can only be evaluated when more about the
`𝕊` is known, for example by knowing the concrete type (e.g. in concrete
translations) or requiring an evaluation function (e.g.
in`ExtensionallyEqual`).