Remove the annotation language from `VariabilityLanguage`
I notices a misconception about the purpose of `F` in
`VariabilityLanguage`. We assumed that `F` is the actual annotation
language in use by `L`, but it's not. It just serves as an argument for
`S` and `S F` determines the actual annotation language. This is fine
for configurations like `Config₂` `F = F → Bool` because it requires `F`
as annotations. However, as soon as we have configurations like
`2Config F = IndexedName F → Bool`, we implicitly require the annotation
language `IndexedName F` not `F`! Hence, the annotation language is not
`F` but something determined by `S F`.

Now, the purpose of `F` seems weird to me. It seems like the semantic of
a language `L` needs to be independent of `F` because it's not the
actual annotation language. Hence, I think we should be able to just
drop `F` and let `S : 𝕊` with `𝕊 = Set` not only determine the
configuration type but implicitly also the annotation type.