One direction of translation preservation is enough
One other half of a preservation proof can be constructed if given one
direction of a translation preservation proof and a proof that the
configuration translations are inverses to each other. However, this is
only useful if the translation functions are extensionally equal to each
other. Hence, in most cases, using this theorem will require function
extensionally.