Try to simplify the module parameters of `Lang.*`
It can be quite annoying to import some of the `Lang.*` modules because
they extensively use module parameters. This led to the creation of
temporary modules to reexport these modules with usable implicit and
explicit parameters. Hence, these module parameters are inlined where
useful.

To further lessen the burden on users a new `Lang.All` module is
created. This should function as a useful public interface to most
languages (`Gruler` is excempt because it doesn't have `(Rose ∞)`
artifacts, or even `Artifact∈ₛVariant`).