Remove `clampAt` and `minFinFromLimit` in favor of `cappedFin`
The interface of `cappedFin` using an implicit parameter which also
doesn't represent the predecessor of the maximum value is very nice to
use. Most uses in the code are already using `cappedFin`.

`minFinFromLimit` also implements the same functionality but is
currently unused and it seems like proofs about it are harder than for
`cappedFin`.