Get rid of `--large-indicies` using higher levels
Agda disables this option by default because it can cause
inconsistencies (especially with `--forced-argument-recursion`, which is
enabled by default). Hence, it's better to disable it as well.

Unfortunately, `MonadWriter` cannot handle different `Level`s for the
Value and the Result type. Hence, `ignore-value` and `return-value` hack
is required to shift the `Level`s as required.

There are still some cases in `IndexedSet` where the `Level`s could be
generalized even more. I know of at least the `Transitive` laws.
However, the standard library doesn't provide a `Trans` definition
allowing different indices as for homogeneous transitivity. We don't
need this generality yet, so I just left it out for now.