Fix the children order of `_⊕_`
Due to the argument order of `_⊙_`, the children got merged in the wrong
order. Hence, the merge order of the children kind of reversed after
each level. Notice that the new argument order agrees with the usual
`List.foldl` argument order.

Example of what changed:
Previously, we had
    ((0 -< rose-leaf 1 ∷ [] >- ∷ []) ⊕ (0 -< rose-leaf 2 ∷ [] >- ∷ []))
  ≡  (0 -< rose-leaf 2 ∷                     rose-leaf 1 ∷ [] >- ∷ [])
and
    ((0 -< 1 -< rose-leaf 2 ∷ [] >- ∷ [] >- ∷ []) ⊕ (0 -< 1 -< rose-leaf 3 ∷ [] >- ∷ [] >- ∷ []))
  ≡  (0 -< 1 -< rose-leaf 2 ∷                                  rose-leaf 3 ∷ [] >- ∷ [] >- ∷ [])
now we have
    ((0 -< rose-leaf 1 ∷ [] >- ∷ []) ⊕ (0 -< rose-leaf 2 ∷ [] >- ∷ []))
  ≡  (0 -< rose-leaf 1 ∷                     rose-leaf 2 ∷ [] >- ∷ [])
and
    ((0 -< 1 -< rose-leaf 2 ∷ [] >- ∷ [] >- ∷ []) ⊕ (0 -< 1 -< rose-leaf 3 ∷ [] >- ∷ [] >- ∷ []))
  ≡  (0 -< 1 -< rose-leaf 2 ∷                                  rose-leaf 3 ∷ [] >- ∷ [] >- ∷ [])