Fix the distant idempotence law of `FeatureAlgebra`
The `FeatureAlgebra` concept is based on the paper "An algebraic
foundation for automatic feature-based program synthesis" by Apel et.
al. which is a little vague and inconsistent with the order of their
composition operator. While discussing the buggy children order fixed in
f3ccb0f895c61152b13be051e450e2dd2a499688, we discovered that the
distant idempotence law is dictated by the order during composition.

Using our current definition of `_⊕_`, we have
    ((0 -< rose-leaf 1 ∷ [] >- ∷ []) ⊕ (0 -< rose-leaf 2 ∷ rose-leaf 3 ∷ [] >- ∷ []))
  ≡  (0 -< rose-leaf 1 ∷                     rose-leaf 2 ∷ rose-leaf 3 ∷ [] >- ∷ [])
which gives rise to the law
  i₁ ⊕ i₂ ⊕ i₁ ≡ i₁ ⊕ i₂

In contrast, with the alternative definition of `_⊕_`, with swapped
composition order, we have
    ((0 -< rose-leaf 1 ∷ [] >- ∷ []) ⊕ (0 -< rose-leaf 2 ∷ rose-leaf 3 ∷ [] >- ∷ []))
  ≡  (0 -< rose-leaf 2 ∷ rose-leaf 3 ∷                     rose-leaf 1 ∷ [] >- ∷ [])
which gives rise to the law
  i₂ ⊕ i₁ ⊕ i₂ ≡ i₁ ⊕ i₂

Although Apel et. al. state the distant idempotence law as
  i₂ ⊕ i₁ ⊕ i₂ ≡ i₁ ⊕ i₂
we decided to use the more intuitive composition order.