Do not use `renaming` to differentiate between languages
Although `renaming` makes the identifiers shorter (especially the
semantics), the suffixes where context dependent (e.g., ₐ, ₙ for ADT,
NADT and ₂, ₙ for 2CC, NCC) and used inconsistently.

Note that The length issue is only really an issue for `VariantList`
which could be named `CaO` as in the paper to make it shorter.