Quantify ≤Size over the artifact type
This makes it easier to use because the artifact type doesn't need to be
applied when invoking `≤Size`. Furthermore, this enables proofs of
`≱Size` to fix a single artifact type, for example the natural numbers,
and automatically have the inhabitants it needs.

The order between the quantifier over `n` and `A` doesn't have a big
impact. On the one hand, the chosen order allows `≱Size` to use
different artifact types for each `n`. However, it doesn't change the
relation inhabitants if they are swapped because there exists a type
with enough elements (i.e., union of all `A` ranging all `n`s) that can
be fixed and then only a subset of the artifacts can be used for a
specific `n`. On the other hand, `≤Size` is a `Set` and, thus, can't be
inspected if the order is changed. This specific order is chosen purely
as it's more convenient for pattern matching (e.g., one less `with`
clause in case of `≤Size`).