Use the new succinctness definition
The designed succinctness definition includes a translatable constraint
that the old definition was missing. This gets rid of the unfortunate
`¬Compiler→¬≤` and `¬Compiler→≤` properties.

A drawback of this new definition is that it breaks transitivity.
Consider some languages L1 and L3 that are complete and a language L2
that is incomplete. There is an expression e in L1 that cannot be
translated to L2. If we have L1 <= L2 and L2 <= L3 we cannot conclude L1
<= L3 because we know nothing about the size of e translated to L3
because we just proved that there exists no translation to L2.

Note that the order of `∀ (A : 𝔸)` and `Σ[ m ∈ ℕ ]` was changed. Due to
parametricity (type parameters cannot be inspected) this does not change
the actual semantics of the definitions. However, it does simplify the
proofs by being friendlier to pattern matching and `with` clauses
avoiding additional helper functions in many cases.