direct VariantList-to-ADT compiler
This compiler is an adapted version of the VariantList-to-VT compiler.
It turned out to be simpler and a better separation of concerns to split
the VariantList-to-VT compiler into two compilers:
VariantList-to-ADT (introduced in this commit) and an ADT-to-VT
compiler (introduced in a previous commit).
Hence, the VariantList-to-ADT compiler and its proofs are a simplified
copy of the VariantList-to-VT compiler.